Solr 9 Updates and Upstream Pull
- Updates for supporting Solr9
- Pulls Upstream code
- Switch to maven
