+++ .gitignore
@@ -9,3 +9,18 @@ build/*
 ci/*
 dist/*
 /bin/
+
+### IntelliJ ###
+.idea/
+*.iml
+
+### Maven ###
+target/
+pom.xml.tag
+pom.xml.releaseBackup
+pom.xml.versionsBackup
+pom.xml.next
+release.properties
+dependency-reduced-pom.xml
+buildNumber.properties
+.mvn/timing.properties
