--- .travis.yml
@@ -1,13 +1,0 @@
-language: java
-script: "ant ci"
-addons:
-  apt:
-    packages:
-      - openjdk-6-jdk
-      - ant
-jdk:
-  # prod, test: Java HostSpot 64-bit Server VM;
-  #     Java(TM) SE Runtime Environment (build 1.6.0_33-b03)
-  - openjdk6
-  - openjdk7
-  - oraclejdk8
