--- build.properties
@@ -1,19 +1,0 @@
-# properties file for CJKFoldingFilter ant build.xml script
-
-java.compat.version = 1.8
-src.dir = src
-build.dir = bin
-dist.dir = dist
-javadoc.dir = docs/javadoc
-
-# --- TEST properties ---
-test.dir = test
-test.lib.dir = ${test.dir}/lib
-test.src.dir = ${test.dir}/src
-test.build.dir = ${build.dir}
-
-# --- Continuous Integration Properties
-ci.dir = ci
-coverage.dir = ${ci.dir}/coverage
-instr.dir = ${ci.dir}/instr
-javadoc.dir = ${ci.dir}/docs
