+++ example/fixtures.json
@@ -1,0 +1,18 @@
+[
+  {
+    "id": "5790412",
+    "cjk_test": "俞平伯点评红楼梦 / 俞平伯著."
+  },
+  {
+    "id": "6435414",
+    "cjk_test": "兩晉南北朝史 / [呂思勉著]"
+  },
+  {
+    "id": "6363404",
+    "cjk_test": "南洋の華僑 / 南洋協會編纂"
+  },
+  {
+    "id": "6750792",
+    "cjk_test": "佛敎學報."
+  }
+]
