+++ src/main/java/edu/stanford/lucene/analysis/CJKFoldingFilter.java
