+++ src/main/java/edu/stanford/lucene/analysis/CJKFoldingFilterFactory.java
