+++ src/main/resources/edu/stanford/lucene/analysis/variantmap.xml
